Update version, set CI, use pyright, remove python2, add typing
