+++ .github/workflows/qa.yml
@@ -1,0 +1,42 @@
+name: QA checks
+
+on:  # yamllint disable-line rule:truthy
+  push:
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    strategy:  # launches
+      fail-fast: false
+      matrix:
+        task:  # yamllint disable rule:indentation
+        - lint
+        - yamllint
+        - check_types
+        - check_fmt
+        - test
+
+    steps:
+    - uses: actions/checkout@v2
+
+    - name: Set up Python 3.9.5
+      uses: actions/setup-python@v2
+      with:
+        python-version: 3.9.5
+
+    - name: Set up cache
+      uses: actions/cache@v2
+      id: cache
+      with:
+        path: /home/runner/.cache/pypoetry/virtualenvs
+        key: venv-${{ hashFiles('**/.github/workflows/qa.yml') }}-${{ hashFiles('**/poetry.lock')
+          }}
+
+    - name: Install dependencies
+      run: |2
+
+        python -m pip install poetry==1.1.8
+        poetry install
+    - name: QA
+      run: poetry run task ${{ matrix['task'] }}
