+++ pyap/__init__.py
@@ -4,4 +4,6 @@
 API hooks
 """
 from .api import parse
-from .utils import (match, findall)
+from .utils import match, findall
+
+__all__ = ["parse", "match", "findall"]
