+++ pyap/source_US/data.py
@@ -17,12 +17,12 @@
 import string
 
 
-'''Numerals from one to nine
+"""Numerals from one to nine
 Note: here and below we use syntax like '[Oo][Nn][Ee]'
 instead of '(one)(?i)' to match 'One' or 'oNe' because
 Python Regexps don't seem to support turning On/Off
 case modes for subcapturing groups.
-'''
+"""
 zero_to_nine = r"""(?:
     [Zz][Ee][Rr][Oo]\ |[Oo][Nn][Ee]\ |[Tt][Ww][Oo]\ |
     [Tt][Hh][Rr][Ee][Ee]\ |[Ff][Oo][Uu][Rr]\ |
@@ -62,7 +62,7 @@ thousand = r"""(?:
     [Tt][Hh][Oo][Uu][Ss][Aa][Nn][Dd]\ 
     )"""
 
-'''
+"""
 Regexp for matching street number.
 Street number can be written 2 ways:
 1) Using letters - "One thousand twenty two"
@@ -70,7 +70,7 @@ Street number can be written 2 ways:
    a) - "1022"
    b) - "85-1190"
    c) - "85 1190"
-'''
+"""
 street_number = r"""(?P<street_number>
                         (?:
                             [Aa][Nn][Dd]\ 
@@ -88,17 +88,19 @@ street_number = r"""(?P<street_number>
                             (?:\ ?\-?\ ?\d{from_to})?\ 
                         )
                     )
-                """.format(thousand=thousand,
-                           hundred=hundred,
-                           zero_to_nine=zero_to_nine,
-                           ten_to_ninety=ten_to_ninety,
-                           from_to='{1,5}')
+                """.format(
+    thousand=thousand,
+    hundred=hundred,
+    zero_to_nine=zero_to_nine,
+    ten_to_ninety=ten_to_ninety,
+    from_to="{1,5}",
+)
 
-'''
+"""
 Regexp for matching street name.
 In example below:
 "Hoover Boulevard": "Hoover" is a street name
-'''
+"""
 street_name = r"""(?P<street_name>
                   [a-zA-Z0-9\ \.]{3,31}  # Seems like the longest US street is
                                          # 'Northeast Kentucky Industrial Parkway'
@@ -128,113 +130,574 @@ post_direction = r"""
 # This list was taken from: https://pe.usps.com/text/pub28/28apc_002.htm
 # Broadway and Lp (abbreviation for Loop) were added to the list
 street_type_list = [
-    'Allee', 'Alley', 'Ally', 'Aly', 'Anex', 'Annex',
-    'Annx', 'Anx', 'Arc', 'Arcade', 'Av', 'Ave',
-    'Aven', 'Avenu', 'Avenue', 'Avn', 'Avnue', 'Bayoo',
-    'Bayou', 'Bch', 'Beach', 'Bend', 'Bg', 'Bgs',
-    'Blf', 'Blfs', 'Bluf', 'Bluff', 'Bluffs', 'Blvd',
-    'Bnd', 'Bot', 'Bottm', 'Bottom', 'Boul', 'Boulevard',
-    'Boulv', 'Br', 'Branch', 'Brdge', 'Brg', 'Bridge',
-    'Brk', 'Brks', 'Brnch', 'Broadway', 'Brook', 'Brooks',
-    'Btm', 'Burg', 'Burgs', 'Byp', 'Bypa', 'Bypas',
-    'Bypass', 'Byps', 'Byu', 'Camp', 'Canyn', 'Canyon',
-    'Cape', 'Causeway', 'Causwa', 'Cen', 'Cent', 'Center',
-    'Centers', 'Centr', 'Centre', 'Cir', 'Circ', 'Circl',
-    'Circle', 'Circles', 'Cirs', 'Clb', 'Clf', 'Clfs',
-    'Cliff', 'Cliffs', 'Club', 'Cmn', 'Cmns', 'Cmp',
-    'Cnter', 'Cntr', 'Cnyn', 'Common', 'Commons', 'Cor',
-    'Corner', 'Corners', 'Cors', 'Course', 'Court', 'Courts',
-    'Cove', 'Coves', 'Cp', 'Cpe', 'Crcl', 'Crcle',
-    'Creek', 'Cres', 'Crescent', 'Crest', 'Crk', 'Crossing',
-    'Crossroad', 'Crossroads', 'Crse', 'Crsent', 'Crsnt', 'Crssng',
-    'Crst', 'Cswy', 'Ct', 'Ctr', 'Ctrs', 'Cts',
-    'Curv', 'Curve', 'Cv', 'Cvs', 'Cyn', 'Dale',
-    'Dam', 'Div', 'Divide', 'Dl', 'Dm', 'Dr',
-    'Driv', 'Drive', 'Drives', 'Drs', 'Drv', 'Dv',
-    'Dvd', 'Est', 'Estate', 'Estates', 'Ests', 'Exp',
-    'Expr', 'Express', 'Expressway', 'Expw', 'Expy', 'Ext',
-    'Extension', 'Extensions', 'Extn', 'Extnsn', 'Exts', 'Fall',
-    'Falls', 'Ferry', 'Field', 'Fields', 'Flat', 'Flats',
-    'Fld', 'Flds', 'Fls', 'Flt', 'Flts', 'Ford',
-    'Fords', 'Forest', 'Forests', 'Forg', 'Forge', 'Forges',
-    'Fork', 'Forks', 'Fort', 'Frd', 'Frds', 'Freeway',
-    'Freewy', 'Frg', 'Frgs', 'Frk', 'Frks', 'Frry',
-    'Frst', 'Frt', 'Frway', 'Frwy', 'Fry', 'Ft',
-    'Fwy', 'Garden', 'Gardens', 'Gardn', 'Gateway', 'Gatewy',
-    'Gatway', 'Gdn', 'Gdns', 'Glen', 'Glens', 'Gln',
-    'Glns', 'Grden', 'Grdn', 'Grdns', 'Green', 'Greens',
-    'Grn', 'Grns', 'Grov', 'Grove', 'Groves', 'Grv',
-    'Grvs', 'Gtway', 'Gtwy', 'Harb', 'Harbor', 'Harbors',
-    'Harbr', 'Haven', 'Hbr', 'Hbrs', 'Heights', 'Highway',
-    'Highwy', 'Hill', 'Hills', 'Hiway', 'Hiwy', 'Hl',
-    'Hllw', 'Hls', 'Hollow', 'Hollows', 'Holw', 'Holws',
-    'Hrbor', 'Ht', 'Hts', 'Hvn', 'Hway', 'Hwy',
-    'Inlet', 'Inlt', 'Is', 'Island', 'Islands', 'Isle',
-    'Isles', 'Islnd', 'Islnds', 'Iss', 'Jct', 'Jction',
-    'Jctn', 'Jctns', 'Jcts', 'Junction', 'Junctions', 'Junctn',
-    'Juncton', 'Key', 'Keys', 'Knl', 'Knls', 'Knol',
-    'Knoll', 'Knolls', 'Ky', 'Kys', 'Lake', 'Lakes',
-    'Land', 'Landing', 'Lane', 'Lck', 'Lcks', 'Ldg',
-    'Ldge', 'Lf', 'Lgt', 'Lgts', 'Light', 'Lights',
-    'Lk', 'Lks', 'Ln', 'Lndg', 'Lndng', 'Loaf',
-    'Lock', 'Locks', 'Lodg', 'Lodge', 'Loop', 'Loops',
-    'Lp', 'Mall', 'Manor', 'Manors', 'Mdw', 'Mdws',
-    'Meadow', 'Meadows', 'Medows', 'Mews', 'Mill', 'Mills',
-    'Mission', 'Missn', 'Ml', 'Mls', 'Mnr', 'Mnrs',
-    'Mnt', 'Mntain', 'Mntn', 'Mntns', 'Motorway', 'Mount',
-    'Mountain', 'Mountains', 'Mountin', 'Msn', 'Mssn', 'Mt',
-    'Mtin', 'Mtn', 'Mtns', 'Mtwy', 'Nck', 'Neck',
-    'Opas', 'Orch', 'Orchard', 'Orchrd', 'Oval', 'Overpass',
-    'Ovl', 'Park', 'Parks', 'Parkway', 'Parkways', 'Parkwy',
-    'Pass', 'Passage', 'Path', 'Paths', 'Pike', 'Pikes',
-    'Pine', 'Pines', 'Pkway', 'Pkwy', 'Pkwys', 'Pky',
-    'Pl', 'Place', 'Plain', 'Plains', 'Plaza', 'Pln',
-    'Plns', 'Plz', 'Plza', 'Pne', 'Pnes', 'Point',
-    'Points', 'Port', 'Ports', 'Pr', 'Prairie', 'Prk',
-    'Prr', 'Prt', 'Prts', 'Psge', 'Pt', 'Pts',
-    'Rad', 'Radial', 'Radiel', 'Radl', 'Ramp', 'Ranch',
-    'Ranches', 'Rapid', 'Rapids', 'Rd', 'Rdg', 'Rdge',
-    'Rdgs', 'Rds', 'Rest', 'Ridge', 'Ridges', 'Riv',
-    'River', 'Rivr', 'Rnch', 'Rnchs', 'Road', 'Roads',
-    'Route', 'Row', 'Rpd', 'Rpds', 'Rst', 'Rte',
-    'Rue', 'Run', 'Rvr', 'Shl', 'Shls', 'Shoal',
-    'Shoals', 'Shoar', 'Shoars', 'Shore', 'Shores', 'Shr',
-    'Shrs', 'Skwy', 'Skyway', 'Smt', 'Spg', 'Spgs',
-    'Spng', 'Spngs', 'Spring', 'Springs', 'Sprng', 'Sprngs',
-    'Spur', 'Spurs', 'Sq', 'Sqr', 'Sqre', 'Sqrs',
-    'Sqs', 'Squ', 'Square', 'Squares', 'St', 'Sta',
-    'Station', 'Statn', 'Stn', 'Str', 'Stra', 'Strav',
-    'Straven', 'Stravenue', 'Stravn', 'Stream', 'Street', 'Streets',
-    'Streme', 'Strm', 'Strt', 'Strvn', 'Strvnue', 'Sts',
-    'Sumit', 'Sumitt', 'Summit', 'Ter', 'Terr', 'Terrace',
-    'Throughway', 'Tpke', 'Trace', 'Traces', 'Track', 'Tracks',
-    'Trafficway', 'Trail', 'Trailer', 'Trails', 'Trak', 'Trce',
-    'Trfy', 'Trk', 'Trks', 'Trl', 'Trlr', 'Trlrs',
-    'Trls', 'Trnpk', 'Trwy', 'Tunel', 'Tunl', 'Tunls',
-    'Tunnel', 'Tunnels', 'Tunnl', 'Turnpike', 'Turnpk', 'Un',
-    'Underpass', 'Union', 'Unions', 'Uns', 'Upas', 'Valley',
-    'Valleys', 'Vally', 'Vdct', 'Via', 'Viadct', 'Viaduct',
-    'View', 'Views', 'Vill', 'Villag', 'Village', 'Villages',
-    'Ville', 'Villg', 'Villiage', 'Vis', 'Vist', 'Vista',
-    'Vl', 'Vlg', 'Vlgs', 'Vlly', 'Vly', 'Vlys',
-    'Vst', 'Vsta', 'Vw', 'Vws', 'Walk', 'Walks',
-    'Wall', 'Way', 'Ways', 'Well', 'Wells', 'Wl',
-    'Wls', 'Wy', 'Xing', 'Xrd', 'Xrds',
+    "Allee",
+    "Alley",
+    "Ally",
+    "Aly",
+    "Anex",
+    "Annex",
+    "Annx",
+    "Anx",
+    "Arc",
+    "Arcade",
+    "Av",
+    "Ave",
+    "Aven",
+    "Avenu",
+    "Avenue",
+    "Avn",
+    "Avnue",
+    "Bayoo",
+    "Bayou",
+    "Bch",
+    "Beach",
+    "Bend",
+    "Bg",
+    "Bgs",
+    "Blf",
+    "Blfs",
+    "Bluf",
+    "Bluff",
+    "Bluffs",
+    "Blvd",
+    "Bnd",
+    "Bot",
+    "Bottm",
+    "Bottom",
+    "Boul",
+    "Boulevard",
+    "Boulv",
+    "Br",
+    "Branch",
+    "Brdge",
+    "Brg",
+    "Bridge",
+    "Brk",
+    "Brks",
+    "Brnch",
+    "Broadway",
+    "Brook",
+    "Brooks",
+    "Btm",
+    "Burg",
+    "Burgs",
+    "Byp",
+    "Bypa",
+    "Bypas",
+    "Bypass",
+    "Byps",
+    "Byu",
+    "Camp",
+    "Canyn",
+    "Canyon",
+    "Cape",
+    "Causeway",
+    "Causwa",
+    "Cen",
+    "Cent",
+    "Center",
+    "Centers",
+    "Centr",
+    "Centre",
+    "Cir",
+    "Circ",
+    "Circl",
+    "Circle",
+    "Circles",
+    "Cirs",
+    "Clb",
+    "Clf",
+    "Clfs",
+    "Cliff",
+    "Cliffs",
+    "Club",
+    "Cmn",
+    "Cmns",
+    "Cmp",
+    "Cnter",
+    "Cntr",
+    "Cnyn",
+    "Common",
+    "Commons",
+    "Cor",
+    "Corner",
+    "Corners",
+    "Cors",
+    "Course",
+    "Court",
+    "Courts",
+    "Cove",
+    "Coves",
+    "Cp",
+    "Cpe",
+    "Crcl",
+    "Crcle",
+    "Creek",
+    "Cres",
+    "Crescent",
+    "Crest",
+    "Crk",
+    "Crossing",
+    "Crossroad",
+    "Crossroads",
+    "Crse",
+    "Crsent",
+    "Crsnt",
+    "Crssng",
+    "Crst",
+    "Cswy",
+    "Ct",
+    "Ctr",
+    "Ctrs",
+    "Cts",
+    "Curv",
+    "Curve",
+    "Cv",
+    "Cvs",
+    "Cyn",
+    "Dale",
+    "Dam",
+    "Div",
+    "Divide",
+    "Dl",
+    "Dm",
+    "Dr",
+    "Driv",
+    "Drive",
+    "Drives",
+    "Drs",
+    "Drv",
+    "Dv",
+    "Dvd",
+    "Est",
+    "Estate",
+    "Estates",
+    "Ests",
+    "Exp",
+    "Expr",
+    "Express",
+    "Expressway",
+    "Expw",
+    "Expy",
+    "Ext",
+    "Extension",
+    "Extensions",
+    "Extn",
+    "Extnsn",
+    "Exts",
+    "Fall",
+    "Falls",
+    "Ferry",
+    "Field",
+    "Fields",
+    "Flat",
+    "Flats",
+    "Fld",
+    "Flds",
+    "Fls",
+    "Flt",
+    "Flts",
+    "Ford",
+    "Fords",
+    "Forest",
+    "Forests",
+    "Forg",
+    "Forge",
+    "Forges",
+    "Fork",
+    "Forks",
+    "Fort",
+    "Frd",
+    "Frds",
+    "Freeway",
+    "Freewy",
+    "Frg",
+    "Frgs",
+    "Frk",
+    "Frks",
+    "Frry",
+    "Frst",
+    "Frt",
+    "Frway",
+    "Frwy",
+    "Fry",
+    "Ft",
+    "Fwy",
+    "Garden",
+    "Gardens",
+    "Gardn",
+    "Gateway",
+    "Gatewy",
+    "Gatway",
+    "Gdn",
+    "Gdns",
+    "Glen",
+    "Glens",
+    "Gln",
+    "Glns",
+    "Grden",
+    "Grdn",
+    "Grdns",
+    "Green",
+    "Greens",
+    "Grn",
+    "Grns",
+    "Grov",
+    "Grove",
+    "Groves",
+    "Grv",
+    "Grvs",
+    "Gtway",
+    "Gtwy",
+    "Harb",
+    "Harbor",
+    "Harbors",
+    "Harbr",
+    "Haven",
+    "Hbr",
+    "Hbrs",
+    "Heights",
+    "Highway",
+    "Highwy",
+    "Hill",
+    "Hills",
+    "Hiway",
+    "Hiwy",
+    "Hl",
+    "Hllw",
+    "Hls",
+    "Hollow",
+    "Hollows",
+    "Holw",
+    "Holws",
+    "Hrbor",
+    "Ht",
+    "Hts",
+    "Hvn",
+    "Hway",
+    "Hwy",
+    "Inlet",
+    "Inlt",
+    "Is",
+    "Island",
+    "Islands",
+    "Isle",
+    "Isles",
+    "Islnd",
+    "Islnds",
+    "Iss",
+    "Jct",
+    "Jction",
+    "Jctn",
+    "Jctns",
+    "Jcts",
+    "Junction",
+    "Junctions",
+    "Junctn",
+    "Juncton",
+    "Key",
+    "Keys",
+    "Knl",
+    "Knls",
+    "Knol",
+    "Knoll",
+    "Knolls",
+    "Ky",
+    "Kys",
+    "Lake",
+    "Lakes",
+    "Land",
+    "Landing",
+    "Lane",
+    "Lck",
+    "Lcks",
+    "Ldg",
+    "Ldge",
+    "Lf",
+    "Lgt",
+    "Lgts",
+    "Light",
+    "Lights",
+    "Lk",
+    "Lks",
+    "Ln",
+    "Lndg",
+    "Lndng",
+    "Loaf",
+    "Lock",
+    "Locks",
+    "Lodg",
+    "Lodge",
+    "Loop",
+    "Loops",
+    "Lp",
+    "Mall",
+    "Manor",
+    "Manors",
+    "Mdw",
+    "Mdws",
+    "Meadow",
+    "Meadows",
+    "Medows",
+    "Mews",
+    "Mill",
+    "Mills",
+    "Mission",
+    "Missn",
+    "Ml",
+    "Mls",
+    "Mnr",
+    "Mnrs",
+    "Mnt",
+    "Mntain",
+    "Mntn",
+    "Mntns",
+    "Motorway",
+    "Mount",
+    "Mountain",
+    "Mountains",
+    "Mountin",
+    "Msn",
+    "Mssn",
+    "Mt",
+    "Mtin",
+    "Mtn",
+    "Mtns",
+    "Mtwy",
+    "Nck",
+    "Neck",
+    "Opas",
+    "Orch",
+    "Orchard",
+    "Orchrd",
+    "Oval",
+    "Overpass",
+    "Ovl",
+    "Park",
+    "Parks",
+    "Parkway",
+    "Parkways",
+    "Parkwy",
+    "Pass",
+    "Passage",
+    "Path",
+    "Paths",
+    "Pike",
+    "Pikes",
+    "Pine",
+    "Pines",
+    "Pkway",
+    "Pkwy",
+    "Pkwys",
+    "Pky",
+    "Pl",
+    "Place",
+    "Plain",
+    "Plains",
+    "Plaza",
+    "Pln",
+    "Plns",
+    "Plz",
+    "Plza",
+    "Pne",
+    "Pnes",
+    "Point",
+    "Points",
+    "Port",
+    "Ports",
+    "Pr",
+    "Prairie",
+    "Prk",
+    "Prr",
+    "Prt",
+    "Prts",
+    "Psge",
+    "Pt",
+    "Pts",
+    "Rad",
+    "Radial",
+    "Radiel",
+    "Radl",
+    "Ramp",
+    "Ranch",
+    "Ranches",
+    "Rapid",
+    "Rapids",
+    "Rd",
+    "Rdg",
+    "Rdge",
+    "Rdgs",
+    "Rds",
+    "Rest",
+    "Ridge",
+    "Ridges",
+    "Riv",
+    "River",
+    "Rivr",
+    "Rnch",
+    "Rnchs",
+    "Road",
+    "Roads",
+    "Route",
+    "Row",
+    "Rpd",
+    "Rpds",
+    "Rst",
+    "Rte",
+    "Rue",
+    "Run",
+    "Rvr",
+    "Shl",
+    "Shls",
+    "Shoal",
+    "Shoals",
+    "Shoar",
+    "Shoars",
+    "Shore",
+    "Shores",
+    "Shr",
+    "Shrs",
+    "Skwy",
+    "Skyway",
+    "Smt",
+    "Spg",
+    "Spgs",
+    "Spng",
+    "Spngs",
+    "Spring",
+    "Springs",
+    "Sprng",
+    "Sprngs",
+    "Spur",
+    "Spurs",
+    "Sq",
+    "Sqr",
+    "Sqre",
+    "Sqrs",
+    "Sqs",
+    "Squ",
+    "Square",
+    "Squares",
+    "St",
+    "Sta",
+    "Station",
+    "Statn",
+    "Stn",
+    "Str",
+    "Stra",
+    "Strav",
+    "Straven",
+    "Stravenue",
+    "Stravn",
+    "Stream",
+    "Street",
+    "Streets",
+    "Streme",
+    "Strm",
+    "Strt",
+    "Strvn",
+    "Strvnue",
+    "Sts",
+    "Sumit",
+    "Sumitt",
+    "Summit",
+    "Ter",
+    "Terr",
+    "Terrace",
+    "Throughway",
+    "Tpke",
+    "Trace",
+    "Traces",
+    "Track",
+    "Tracks",
+    "Trafficway",
+    "Trail",
+    "Trailer",
+    "Trails",
+    "Trak",
+    "Trce",
+    "Trfy",
+    "Trk",
+    "Trks",
+    "Trl",
+    "Trlr",
+    "Trlrs",
+    "Trls",
+    "Trnpk",
+    "Trwy",
+    "Tunel",
+    "Tunl",
+    "Tunls",
+    "Tunnel",
+    "Tunnels",
+    "Tunnl",
+    "Turnpike",
+    "Turnpk",
+    "Un",
+    "Underpass",
+    "Union",
+    "Unions",
+    "Uns",
+    "Upas",
+    "Valley",
+    "Valleys",
+    "Vally",
+    "Vdct",
+    "Via",
+    "Viadct",
+    "Viaduct",
+    "View",
+    "Views",
+    "Vill",
+    "Villag",
+    "Village",
+    "Villages",
+    "Ville",
+    "Villg",
+    "Villiage",
+    "Vis",
+    "Vist",
+    "Vista",
+    "Vl",
+    "Vlg",
+    "Vlgs",
+    "Vlly",
+    "Vly",
+    "Vlys",
+    "Vst",
+    "Vsta",
+    "Vw",
+    "Vws",
+    "Walk",
+    "Walks",
+    "Wall",
+    "Way",
+    "Ways",
+    "Well",
+    "Wells",
+    "Wl",
+    "Wls",
+    "Wy",
+    "Xing",
+    "Xrd",
+    "Xrds",
 ]
 
 
 def street_type_list_to_regex(street_type_list):
     """Converts a list of street types into a regex"""
-    street_types = '|'.join(set(street_type_list)).lower()
+    street_types = "|".join(set(street_type_list)).lower()
     for letter in string.ascii_lowercase:
-        street_types = street_types.replace(letter, '[{upper}{lower}]'.format(upper=letter.upper(), lower=letter))
+        street_types = street_types.replace(
+            letter, "[{upper}{lower}]".format(upper=letter.upper(), lower=letter)
+        )
 
     # Use \b to check that there are word boundaries before and after the street type
     # Optionally match zero to two of " ", ",", or "." after the street name
-    street_types = street_types.replace('|', r'\b{div}|\b')
-    street_types = r'\b' + street_types + r'\b{div}'
+    street_types = street_types.replace("|", r"\b{div}|\b")
+    street_types = r"\b" + street_types + r"\b{div}"
     return street_types.format(
-        div=r'[\.\ ,]{0,2}',
+        div=r"[\.\ ,]{0,2}",
     )
 
 
@@ -250,7 +713,7 @@ street_type = r"""
                 )?
             )
 """.format(
-    route_symbols='{0,3}',
+    route_symbols="{0,3}",
     street_types=street_type_list_to_regex(street_type_list),
 )
 
@@ -291,11 +754,12 @@ building = r"""
                 )
                 \ ?
             )
-            """.format(thousand=thousand,
-                       hundred=hundred,
-                       zero_to_nine=zero_to_nine,
-                       ten_to_ninety=ten_to_ninety,
-                       )
+            """.format(
+    thousand=thousand,
+    hundred=hundred,
+    zero_to_nine=zero_to_nine,
+    ten_to_ninety=ten_to_ninety,
+)
 
 occupancy = r"""
             (?P<occupancy>
@@ -341,15 +805,16 @@ full_street = r"""
             {occupancy}?\,?\ ?
             {po_box}?
         )
-    )""".format(street_number=street_number,
-                street_name=street_name,
-                street_type=street_type,
-                post_direction=post_direction,
-                floor=floor,
-                building=building,
-                occupancy=occupancy,
-                po_box=po_box,
-                )
+    )""".format(
+    street_number=street_number,
+    street_name=street_name,
+    street_type=street_type,
+    post_direction=post_direction,
+    floor=floor,
+    building=building,
+    occupancy=occupancy,
+    po_box=po_box,
+)
 
 # region1 is actually a "state"
 region1 = r"""
@@ -459,7 +924,7 @@ full_address = r"""
                 )
                 """.format(
     full_street=full_street,
-    div=r'[\, ]{,2}',
+    div=r"[\, ]{,2}",
     city=city,
     region1=region1,
     country=country,
